chg: Moved StrToCamel Into kstring Package
BREAKING CHANGE
