--- strings.go
@@ -1,48 +1,0 @@
-package stdlib
-
-import (
-	"bytes"
-	"fmt"
-	"regexp"
-)
-
-// StrToCamel convert a dash/underscore separated string to camel case. Works great on pretty/vanity URLs.
-//
-// camel case - a typographical convention in which an initial capital is
-// used for the first letter of a word forming the second element of a
-// closed compound, e.g. iPhone, eBay, PayPal.
-//
-// Pascal case -- or PascalCase - is a programming naming convention where
-// the first letter of each compound word in a variable is capitalized.
-func StrToCamel(subject, separate string, pascal bool) (string, error) {
-	if len(subject) < 1 {
-		return subject, fmt.Errorf("there is nothing to do, the string is empty %q", subject)
-	}
-
-	// Look for the separator, when found return it and the following char if any.
-	re, err1 := regexp.Compile("(?:^|" + separate + ")(.?)")
-
-	if err1 != nil {
-		return "", err1
-	}
-
-	sb := []byte(separate)
-	rVal := re.ReplaceAllFunc([]byte(subject), func(match []byte) []byte {
-		if len(match) == 2 { // match the separator plus 1 letter.
-			b := bytes.ToUpper(match)[1]
-			return []byte{b}
-		}
-
-		if sb[0] != match[0] { // match the first letter
-			if pascal {
-				return bytes.ToUpper(match)
-			} else {
-				return bytes.ToLower(match)
-			}
-		}
-
-		return []byte{}
-	})
-
-	return string(rVal), nil
-}
